backend/requirements.txt
accelerate==1.11.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bidict==0.23.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
feedparser==6.0.12
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
oauthlib==3.3.1
orjson==3.11.4
packaging==25.0
psutil==7.1.3
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
safetensors==0.6.2
sentry-sdk==2.44.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
tweepy==4.16.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn[standard]
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.3.1
